+++ src/app/_components/sidebar/trand-ranking.tsx
@@ -11,7 +11,6 @@ export default function TrendRanking() {
 
   const handleClick = () => {
     router.push(ROUTES.SEARCH);
-
   };
 
   const toggleDropdown = () => {
@@ -22,12 +21,9 @@ export default function TrendRanking() {
   const dummyTimestamp = "2024.04.20 17:32";
 
   return (
-
     <div className="flex w-75 flex-col border-y border-gray-200">
       <div className="flex items-center justify-between px-5 py-2.5">
-        <div
-          className="group body1 flex w-58 items-center gap-3 font-medium text-black"
-        >
+        <div className="group body1 flex w-58 items-center gap-3 font-medium text-black">
           <span>1</span>
           <span className="cursor-pointer group-hover:underline">
             {trends[0]}
